Move reducer out off App.tsx
