+++ src/project/ProjectGrid.stories.tsx
@@ -2,7 +2,7 @@ import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { manipulateData } from '../util/projectData';
 import { ProjectGrid } from './ProjectGrid';
 import mockData from "../api/data.json"
-import { handleProjectLoaded } from '../App';
+import { handleProjectLoaded } from '../actions/root.reducer';;
 import { action } from '@storybook/addon-actions';
 import { TogglProjectResponse } from '../types';
 import { Map, Set } from 'immutable';
